update plots mmodify
